split tests out of file so sbt can find them
